requirements_talking_video.txt
# Core dependencies for Talking Video Generator
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
TTS>=0.15.0
opencv-python>=4.6.0
pillow>=9.0.0
numpy>=1.21.0
gdown>=4.5.0

# Additional dependencies (installed by SadTalker)
face-alignment>=1.3.5
imageio>=2.19.0
imageio-ffmpeg>=0.4.7
librosa>=0.9.0
numba>=0.56.0
resampy>=0.3.0
scipy>=1.9.0
scikit-image>=0.19.0
tqdm>=4.64.0
yacs>=0.1.8
